+++ src/types/public-api/mappings.ts
@@ -42,6 +42,17 @@ export type chiftOperations = {
     updateSyncToConsumer: 'syncs_update_synctoconsumer';
     enableSyncConsumer: 'syncs_enable_syncconsumer';
     updateFlowToConsumer: 'syncs_update_flowtoconsumer';
+    getSyncExecutions: 'syncs_get_sync_executions';
+    disableSyncConsumer: 'syncs_disable_syncconsumer';
+
+    // Datalayer operations
+    enableDatalayer: 'datalayer_enable';
+    refreshDatalayer: 'datalayer_refresh';
+    disableDatalayer: 'datalayer_disable';
+
+    // Datalab operations
+    getCubeSchemas: 'datalab_get_cube_schemas';
+    queryDb: 'datalab_query_db';
 
     // Datastore operations
     getDatastores: 'datastores_get_datastores';
@@ -113,6 +124,8 @@ export type chiftOperations = {
     createFinancialEntries: 'accounting_create_financial_entries';
     getOutstandings: 'accounting_get_outstandings';
     createInvoicePayment: 'accounting_create_invoice_payment';
+    getFolder: 'accounting_get_folder';
+    getPartnerContacts: 'accounting_get_partner_contacts';
 
     // POS operations
     getOrders: 'pos_get_orders';
@@ -130,6 +143,7 @@ export type chiftOperations = {
     getAccountingCategories: 'pos_get_accounting_categories';
     getClosure: 'pos_get_closure';
     getObjectives: 'pos_get_objectives';
+    getPOSTaxes: 'pos_get_taxes';
 
     // eCommerce operations
     getCommerceCustomers: 'ecommerce_get_customers';
@@ -180,6 +194,8 @@ export type chiftOperations = {
     getPaymentPayments: 'payment_get_payments';
     getPayment: 'payment_get_payment';
     getRefunds: 'payment_get_refunds';
+    getPaymentLocations: 'payment_get_locations';
+    getPayouts: 'payment_get_payouts';
 
     // PMS operations
     getPMSOrders: 'pms_get_orders';
@@ -192,4 +208,5 @@ export type chiftOperations = {
     getPMSAccountingCategories: 'pms_get_accounting_categories';
     getPMSClosure: 'pms_get_closure';
     getPMSTaxes: 'pms_get_taxes';
+    getPMSAccountingTransactions: 'pms_get_accounting_transactions';
 };